+++ tests/testthat/test-brace_linter.R
@@ -1,4 +1,7 @@
-test_that("brace_linter lints closed braces correctly", {
+test_that("brace_linter lints braces correctly", {
+  open_curly_msg <- rex::rex(
+    "Opening curly braces should never go on their own line and should always be followed by a new line."
+  )
   closed_curly_msg <- rex::rex(paste(
     "Closing curly-braces should always be on their own line,",
     "unless they are followed by an else."
@@ -8,7 +11,7 @@ test_that("brace_linter lints closed braces correctly", {
   expect_lint("blah", NULL, linter)
   expect_lint("a <- function() {\n}", NULL, linter)
 
-  expect_lint("a <- function() { 1 }", closed_curly_msg, linter)
+  expect_lint("a <- function() { 1 }", list(open_curly_msg, closed_curly_msg), linter)
   # allowed by allow_single_line
   expect_lint("a <- function() { 1 }", NULL, brace_linter(allow_single_line = TRUE))
 
@@ -47,7 +50,7 @@ test_that("brace_linter lints closed braces correctly", {
   )
 
   # }) is allowed
-  expect_lint("eval(bquote({...}))", NULL, linter)
+  expect_lint("eval(bquote({\n...\n}))", NULL, linter)
 
   # }, is allowed
   expect_lint(
@@ -77,9 +80,39 @@ test_that("brace_linter lints closed braces correctly", {
     linter
   )
 
-  # }} is allowed
+  # {{ }} is allowed
   expect_lint("{{ x }}", NULL, linter)
+
+  expect_lint(
+    trim_some("
+      pkg_name <- function(path = find_package()) {
+        if (is.null(path)) {
+          return(NULL)
+        } else {
+          read.dcf(file.path(path, \"DESCRIPTION\"), fields = \"Package\")[1]
+        }
+      }
+    "),
+    NULL,
+    linter
+  )
+
+  expect_lint("a <- function()\n{\n  1 \n}", open_curly_msg, linter)
+  expect_lint("a <- function()\n    {\n  1 \n}", open_curly_msg, linter)
+  expect_lint("a <- function()\n\t{\n  1 \n}", open_curly_msg, linter)
+
+  # trailing comments are allowed
+  expect_lint(
+    trim_some('
+      if ("P" != "NP") { # what most people expect
+        print("Cryptomania is possible")
+      }
+    '),
+    NULL,
+    linter
+  )
 })
+
 
 test_that("brace_linter lints else correctly", {
   linter <- brace_linter()
@@ -145,8 +178,8 @@ test_that("brace_linter lints function expressions correctly", {
   )
 })
 
-test_that("braces_linter lints if/else matching braces correctly", {
-  linter <- braces_linter()
+test_that("brace_linter lints if/else matching braces correctly", {
+  linter <- brace_linter()
   expect_lint("if (TRUE) 1 else 2", NULL, linter)
   expect_lint("if (TRUE) 1", NULL, linter)